LayoutTests/accessibility/hit-test-input-auto-fill-button.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
<script src="resources/shouldBeAccessibleByCursor.js"></script>
</head>
<body>
<div id="description"></div>
<input type="password" id="password">
<div id="console"></div>
<script>
description("This tests that the AutoFill button is hit testable.");

if (!window.internals || !window.accessibilityController)
    testFailed("You must run this test in the test tool and have implemented window.internals and window.accessibilityController.");
else {
    internals.setShowAutoFillButton(document.getElementById("password"), true);

    shouldBeAccessibleByCursor(accessibilityController.accessibleElementById("password").childAtIndex(0), "AutoFill button");
}
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
